fix logout fragment to activity
