+++ app/src/main/res/layout/activity_logout.xml
@@ -5,10 +5,9 @@
     <data>
 
         <variable
-            name="logoutFragment"
-            type="com.thinlineit.ctrlf.main.LogoutFragment" />
+            name="LogoutActivity"
+            type="com.thinlineit.ctrlf.main.LogoutActivity" />
     </data>
-
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
         android:layout_height="match_parent"
@@ -31,7 +30,7 @@
                 android:adjustViewBounds="true"
                 android:src="@drawable/ic_logo_0817"
                 app:layout_constraintStart_toStartOf="parent"
-                app:layout_constraintTop_toTopOf="parent"></ImageView>
+                app:layout_constraintTop_toTopOf="parent"/>
         </androidx.appcompat.widget.Toolbar>
 
         <androidx.cardview.widget.CardView
@@ -59,7 +58,7 @@
                     app:layout_constraintBottom_toTopOf="@+id/logoutPhraseText"
                     app:layout_constraintEnd_toEndOf="parent"
                     app:layout_constraintStart_toStartOf="parent"
-                    app:layout_constraintTop_toTopOf="parent"></TextView>
+                    app:layout_constraintTop_toTopOf="parent"/>
 
                 <TextView
                     android:id="@+id/logoutPhraseText"
@@ -72,7 +71,7 @@
                     android:textColor="@color/black"
                     app:layout_constraintBottom_toBottomOf="parent"
                     app:layout_constraintEnd_toEndOf="parent"
-                    app:layout_constraintStart_toStartOf="parent"></TextView>
+                    app:layout_constraintStart_toStartOf="parent"/>
 
             </androidx.constraintlayout.widget.ConstraintLayout>
         </androidx.cardview.widget.CardView>
@@ -89,7 +88,7 @@
             app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"></Button>
+            app:layout_constraintTop_toTopOf="parent"/>
 
         <Button
             android:id="@+id/cancelBtn"
@@ -102,7 +101,7 @@
             android:text="@string/cancel"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@+id/logoutBtn"></Button>
+            app:layout_constraintTop_toBottomOf="@+id/logoutBtn"/>
 
     </androidx.constraintlayout.widget.ConstraintLayout>
