+++ app/src/main/res/layout/fragment_page_title.xml
@@ -94,7 +94,6 @@
             app:data="@{pageViewModel.topicInfo}"
             app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
             app:layout_constraintTop_toBottomOf="@+id/topicInfo"
-            app:layout_constraintBottom_toTopOf="@id/addPageBtn"
             tools:listitem="@layout/list_item_page_title" />
 
         <Button
